Split redux store into separate files and removed extra files in the components folder
