+++ client/components/AllChannels.js
@@ -21,7 +21,7 @@ class AllChannels extends Component {
 }
 
 const mapStateToProps = state => ({
-  channels: state.channels
+  channels: state.channelsObj.channels
 })
 
 const mapDispatchToProps = dispatch => ({
